+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeNetworking/types.ts
@@ -4,6 +4,9 @@ export type IPTypes =
   | 'IPv4 – Reserved (private)'
   | 'IPv4 – Reserved (public)'
   | 'IPv4 – Shared'
+  | 'IPv4 – VPC – Range'
+  | 'IPv4 – VPC'
   | 'IPv6 – Link Local'
   | 'IPv6 – Range'
-  | 'IPv6 – SLAAC';
+  | 'IPv6 – SLAAC'
+  | 'VPC IPv4 – NAT';